+++ test/src/turtle/turtle_encoder_test.dart
@@ -857,7 +857,7 @@ void main() {
 
       // Act
       final result = encoder.convert(graph);
-      print(result);
+      //print(result);
       // Assert
       expect(
         result.trim(),